Move code from CLI lib to webcomponent builder
